+++ AddOns/DBM-Challenges/CHANGELOG.md
@@ -1,9 +1,34 @@
 # <DBM Mod> Challenges
 
-## [r168](https://github.com/DeadlyBossMods/DBM-Challenges/tree/r168) (2025-04-22)
-[Full Changelog](https://github.com/DeadlyBossMods/DBM-Challenges/compare/r167...r168) [Previous Releases](https://github.com/DeadlyBossMods/DBM-Challenges/releases)
+## [r170](https://github.com/DeadlyBossMods/DBM-Challenges/tree/r170) (2025-05-29)
+[Full Changelog](https://github.com/DeadlyBossMods/DBM-Challenges/compare/r169...r170) [Previous Releases](https://github.com/DeadlyBossMods/DBM-Challenges/releases)
 
-- apparently i never remembered to bump tocs  
-- populate TWW variant creature and encounter IDs  
-- Enable visions loading in visions revisited  
-- Because DBT can't get it's shit together, all 77 (and counting) mods that do this just need to work around core bug  
+- Vision update:  
+     - Added nameplate timer for touch of the abyss  
+     - Explosive leap nameplate timer will now glow as high priority nameplate timer)  
+     - Surging Fist changed to hybrid timer (so it shows on nameplate in addition to showing regular timer  
+     - Decimator changed to hybrid timer (so it shows on nameplate in addition to showing regular timer  
+     - Toxic Volley changed to hybrid timer (so it shows on nameplate in addition to showing regular timer  
+     - Toxic breath changed to hybrid timer (so it shows on nameplate in addition to showing regular timer  
+     - Vez'okk will now show initial timers on engage  
+     - Decimator Shiq'voth will now show initial timers on engage  
+     - Aqir Mindhunter will now show initial timers on engage  
+- fixed bug causing war stomp timer not to update  
+- Visions update:  
+     - Added alert for when Gamon spawns  
+     - Added an alert and nameplate timer for gamons dangerous war stomp  
+     - Changed Dark forge alert to be a pre cast warning to dodge frontal now that it doesn't follow/track in TWW (like it did in bfa)  
+     - Added darkforce hybrid timer (cd and nameplate)  
+     - Added Orb of annihilation hybrid timer  
+     - Added explosive Leap hybrid timer  
+     - Added desperate Retching hybrid timer  
+     - Added maddening Roar hybrid timer  
+- scrap haunting shadows stuff. blizz broke ability to detect it using nameplates in TWW  
+    Changed horrifying shout and abyss casts to modern priority nameplate cast timer icons  
+- use correct gossip ID  
+- Visions Update:  
+    Attempt to fix sanity alerts  
+    Start data collection for initial nameplate timers  
+    Added dark smash alert and nameplate timer  
+    Added blade flourish nameplate timer  
+- fixes to ordinance timer